Remove closed flag on halfArea and make sphericalRingArea insensitive to winding order.
Mike,

Your code is great as always. Here are mainly three suggestions that you should, of course, feel free to ignore if not useful:

1) The closed flag is not needed in halfArea and the code can be modified with this flag effectively being false.

When the flag is set to false but the ring is actually closed, the first evaluation of the loop compares two identical points (λ0 === λ1 && φ0 === φ1).
We then have u = 1 (Pythagorean trigonometric identity) and v = 0 whence Math.atan2( v, u ) = 0 (all we need is that u > 0, per atan2 spec).
This adds no incremental area and the flag value therefore does not matter for rings that are closed.

Always setting the closed flag to false has the advantage that open rings (such as triangles during filtering) are automatically completed (closed).

2) planarRingArea is not sensitive to the winding of a ring, but sphericalRingArea is. When interior === false but winding is counter-clockwise, the returned area is that of the sphere with the ring as its hole. I would argue that this is a bad (and undocumented) inconsistency with planarRingArea; the expansion of the function signature to accept a "interior" flag is another symptom of this problem.

If this pull request were accepted, uses of the new sphericalRingArea should also be updated to remove the "interior" flag. 

3) A remaining inconsistency between planarRingArea and sphericalRingArea concerns the handling of rings straddling the anti-meridian. Trying to take the viewpoint of your other libraries, I think sphericalRingArea handles these rings correctly while planarRingArea interprets the line [[-10, 0],[180, 0]] as "the long way around". I have not filed a corresponding edit for planarRingArea.

4) The successor of halfArea could be exported for re-use beyond small-area filtering and line simplification (reflected in the pull request, but not essential). I also recall seeing you use Kahan compensation elsewhere and this halfArea-successor would be a good candidate for its use as well.

Best wishes,

M.